# openapi.yaml
openapi: 3.0.3
info:
  title: MintaiDialect API
  description: 闽台方言语音识别与数字嘉庚对话接口，目前对话可以使用普通话和闽南话作为输入，ASR和TTS暂时只支持闽南语音频和文本互转
  version: 1.0.0
servers:
  - url: http://115.190.116.189
security:
  - BearerAuth: ["admin-token-PbwEQUSPXLY"]
paths:
  /api/digital-jiageng/chat:
    post:
      summary: 与数字嘉庚对话
      description: 上传音频文件与数字嘉庚进行对话。支持可选字幕返回（show_subtitles=true 时返回）。
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/JiagengChatRequest"
      responses:
        "200":
          description: 对话成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaseResponseJiageng"
              example:
                success: true
                message: "操作成功"
                data:
                  response_audio_url: "/uploads/jiageng.wav"
                  subtitles:
                    - text: "您好，我是嘉庚"
                      start_time: 0.0
                      end_time: 0.8
                    - text: "今天"
                      start_time: 0.8
                      end_time: 1.5
        "400":
          description: 缺少必要输入或参数不合法（如音频格式不支持、文件过大）
        "502":
          description: ASR识别失败或LLM/TTS 生成失败
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  # 可选，表示 token 格式
  schemas:
    JiagengChatRequest:
      type: object
      required:
        - audio_file
      properties:
        audio_file:
          type: string
          format: binary
          description: 前端录音文件（音频格式），文件大小限制10MB
        input_language:
          type: string
          enum: [minnan, mandarin, hakka, taiwanese]
          default: minnan
          description: 输入的语种（minnan选项同时可以识别普通话）
        output_language:
          type: string
          enum: [minnan, mandarin, hakka, taiwanese]
          default: minnan
          description: 数字人回复的方言语种（目前只支持闽南话）
        speaking_speed:
          type: number
          format: float
          default: 1.0
          description: 数字人语速
          minimum: 0.5
          maximum: 2.0
        show_subtitles:
          type: boolean
          default: true
          description: 是否返回字幕
    BaseResponseJiageng:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "操作成功"
        data:
          type: object
          properties:
            response_audio_url:
              type: string
              example: "/uploads/jiageng.wav"
            subtitles:
              type: array
              items:
                type: object
                properties:
                  text:
                    type: string
                    example: "您好，我是嘉庚"
                  start_time:
                    type: number
                    format: float
                    example: 0.0
                  end_time:
                    type: number
                    format: float
                    example: 0.8
...